Feat: Crud request

+++ src/main/java/com/fiap/persistence/ecommerce/infrastructure/repository/product/entity/ProductEntity.java
@@ -4,6 +4,8 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
+
 @Data
 @Document
 public class ProductEntity {
@@ -12,6 +14,6 @@ public class ProductEntity {
     private String productId;
     private String name;
     private String quantity;
-    private Long actualUnitValue;
+    private BigDecimal actualUnitValue;
 
 }
